requirements.txt
streamlit==1.31.0
python-docx==1.0.0
pandas==2.0.3
numpy==1.24.3 